Move FetchParent to its own module
