+++ spinetoolbox/spine_db_editor/mvcmodels/compound_parameter_models.py
@@ -19,7 +19,8 @@ These models concatenate several 'single' models and one 'empty' model.
 from PySide2.QtCore import Qt, Slot, QTimer, QModelIndex
 from PySide2.QtGui import QFont
 from spinedb_api.parameter_value import join_value_and_type
-from ...helpers import rows_to_row_count_tuples, parameter_identifier, FetchParent
+from ...helpers import rows_to_row_count_tuples, parameter_identifier
+from ...fetch_parent import FetchParent
 from ..widgets.custom_menus import ParameterViewFilterMenu
 from ...mvcmodels.compound_table_model import CompoundWithEmptyTableModel
 from .empty_parameter_models import (
