+++ spinetoolbox/spine_db_editor/mvcmodels/multi_db_tree_item.py
@@ -18,7 +18,8 @@ Base classes to represent items from multiple databases in a tree.
 from operator import attrgetter
 
 from PySide2.QtCore import Qt
-from ...helpers import rows_to_row_count_tuples, bisect_chunks, FetchParent
+from ...helpers import rows_to_row_count_tuples, bisect_chunks
+from ...fetch_parent import FetchParent
 from ...mvcmodels.minimal_tree_model import TreeItem
