+++ spinetoolbox/spine_db_editor/mvcmodels/tree_item_utility.py
@@ -19,7 +19,8 @@ A tree model for parameter_value lists.
 from PySide2.QtCore import Qt
 from PySide2.QtGui import QBrush, QFont, QIcon, QGuiApplication
 from spinetoolbox.mvcmodels.minimal_tree_model import TreeItem
-from spinetoolbox.helpers import CharIconEngine, FlexibleFetchParent, bisect_chunks
+from spinetoolbox.helpers import CharIconEngine, bisect_chunks
+from spinetoolbox.fetch_parent import FlexibleFetchParent
 
 
 class StandardTreeItem(TreeItem):
