+++ spinetoolbox/spine_db_editor/widgets/graph_view_mixin.py
@@ -23,7 +23,8 @@ from PySide2.QtCore import Slot, QTimer, QThreadPool
 from PySide2.QtWidgets import QHBoxLayout
 from spinedb_api import from_database
 from ...widgets.custom_qgraphicsscene import CustomGraphicsScene
-from ...helpers import get_save_file_name_in_last_dir, FlexibleFetchParent
+from ...helpers import get_save_file_name_in_last_dir
+from ...fetch_parent import FlexibleFetchParent
 from ..graphics_items import (
     EntityItem,
     ObjectItem,
